// src/purchases/types/purchase-result.type.ts
export type PurchasesResult = {
  purchase_id: string;
  purchase_date: string;
  total_products_purchased: number;
  total_stocks_added: number;
  total_invested: number;
  created_at: string;
  last_update: string;
};

type PurchasesQueryResult = {
  purchase_id: string;
  purchase_date: string;
  total_products_purchased: number;
  total_stocks_added: number;
  total_invested: number;
  created_at: string;
  last_update: string;
};